Changed how mapping arb to eng works
In this commit I decided to use Maps to map arabic words with english words.

- Identifiers -> renamed to keywords
- WebAPI -> Map for WebAPI words with arabic words, in progress
- TODO: Change punctuations.ts to use maps
- TODO: Continue mapping out the WebAPI
- TODO: Create a test script to ensure that there are no duplicate identifiers

The most difficult thing about this is the fact that in english we can write camelCase and PascalCase words if we want. Arabic letters are joined so that means there needs to be a single word that represents a a multi word identifier. For example, "اسمع" which translates to "listen" represents addEventListener. LOL. That can either be a great advantage or disadvantage of GahwaScript, depending on how you choose to look at it.

"عليك بتقوى الله ان كنت غافلا، يأتيك بالأرزاق من حيث لا تدري"