Compiler-v2 -> Works!
I've written a new compiler that is functionally nearly identical to the previous compiler, with some minor modifications.

The good news is it works, sort of. there's still the headache of chained props and methods, etc. but hey, at least it works.

I've included a basic HTML file with some bootstrap 5 (which supports RTL languages now, sweet) that has a basic text editor and a button to run everything.

I think it would be good to eliminate the need to open the console by the user and that's something to work on next but I'm not sure how to implement it.

Great stuff! 😁